% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_rows_cols.R
\name{array_drop}
\alias{array_drop}
\title{Drop parts of the dimensions of an array}
\usage{
array_drop(x, ...)
}
\arguments{
\item{x}{A matrix or an array.}

\item{\ldots}{Exactly K integer arguments where
K = \code{length(dim(x))}.
The exception is for matrices, where K == 1 is accepted if the
the argument is named either \code{rows} or \code{cols}.}

\item{A}{subset of \code{x} with \code{length(dim(x))} being preserved.}
}
\description{
Drop parts of the dimensions of an array
}
\examples{
x <- as.list(1:12)
dim(x) <- c(4, 3)
dimnames(x) <- list(letters[1:4], LETTERS[1:3])
y <- as.listenv(x)
x <- array_drop(x, cols = 2)
print(x)
y <- array_drop(y, cols = 2)
print(as.list(y))

x <- as.list(1:12)
dim(x) <- c(2, 2, 3)
dimnames(x) <- list(c("u", "v"), c("a", "b"), c("A", "B", "C"))
print(x)
x <- array_drop(x, 1, NULL, 3)
print(x)

}
