% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listenv.R
\name{as.list.listenv}
\alias{as.list.listenv}
\title{List representation of a list environment}
\usage{
\method{as.list}{listenv}(x, ...)
}
\arguments{
\item{x}{A list environment.}

\item{...}{Not used.}
}
\value{
A list.
}
\description{
List representation of a list environment
}
\keyword{internal}

