% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listenv.R
\name{$.listenv}
\alias{$.listenv}
\alias{[[.listenv}
\title{Get elements of list environment}
\usage{
\method{$}{listenv}(x, name)
}
\arguments{
\item{x}{A list environment.}

\item{name}{The name or index of the element to retrieve.}
}
\value{
The value of an element or NULL if the element does not exist
}
\description{
Get elements of list environment
}
\keyword{internal}
