% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listenv.R
\name{$<-.listenv}
\alias{$<-.listenv}
\alias{[[<-.listenv}
\alias{[<-.listenv}
\title{Set an element of list environment}
\usage{
\method{$}{listenv}(x, name) <- value
}
\arguments{
\item{x}{A list environment.}

\item{name}{Name or index of element}

\item{value}{The value to assign to the element}
}
\description{
Set an element of list environment
}
\keyword{internal}
