% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listenv.R
\name{length.listenv}
\alias{length.listenv}
\alias{lengths.listenv}
\title{Number of elements in list environment}
\usage{
\method{length}{listenv}(x)
}
\arguments{
\item{x}{A list environment.}
}
\description{
Number of elements in list environment
}
\keyword{internal}

