% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listenv.R
\name{map}
\alias{map}
\alias{map.listenv}
\title{Variable name map for elements of list environment}
\usage{
map(x, ...)
}
\arguments{
\item{x}{A list environment.}
}
\value{
The a named character vector
}
\description{
Variable name map for elements of list environment
}
\keyword{internal}

