% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listenv.R
\name{mapping}
\alias{mapping}
\alias{mapping.listenv}
\alias{map}
\title{Name map for elements of list environment}
\usage{
mapping(x, ...)

map(x, ...)
}
\arguments{
\item{x}{A list environment.}
}
\value{
A named character vector
}
\description{
Name map for elements of list environment
}
\keyword{internal}
