% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listenv.R
\name{names.listenv}
\alias{names.listenv}
\alias{names<-.listenv}
\title{Names of elements in list environment}
\usage{
\method{names}{listenv}(x)
}
\arguments{
\item{x}{A list environment.}
}
\description{
Names of elements in list environment
}
\keyword{internal}

