% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undim.R
\name{undim}
\alias{undim}
\alias{undim.default}
\alias{undim.listenv}
\title{Removes the dimension of an object}
\usage{
undim(x, ...)
}
\arguments{
\item{x}{An object with or without dimensions}

\item{...}{Not used.}
}
\value{
The object with the dimension attribute removed.
}
\description{
Removes the dimension of an object
}
\details{
This function does \code{attr(x, "dim") <- NULL}, which automatically also does
\code{attr(x, "dimnames") <- NULL}.
However, other attributes such as names attributes are preserved,
which is not the case if one does \code{dim(x) <- NULL}.
}
\keyword{internal}
