% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aperm.R
\name{aperm.listenv}
\alias{aperm.listenv}
\title{Transpose a 'listenv' array by permuting its dimensions}
\usage{
\method{aperm}{listenv}(a, perm, ...)
}
\arguments{
\item{a}{(listenv) The list environment to be transposed}

\item{perm}{(integer vector) An index vector of length \code{dim(a)}}

\item{\ldots}{Additional arguments passed to \code{\link[base:aperm]{base::aperm()}}.}
}
\value{
Returns a list environment with permuted dimensions
}
\description{
Transpose a 'listenv' array by permuting its dimensions
}
\seealso{
This works \code{\link[base:aperm]{base::aperm()}}.
}
