% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listenv.R
\name{as.list.listenv}
\alias{as.list.listenv}
\title{List representation of a list environment}
\usage{
\method{as.list}{listenv}(x, all.names = TRUE, sorted = FALSE, ...)
}
\arguments{
\item{x}{A list environment.}

\item{all.names}{If \code{TRUE}, variable names starting with
a period are included, otherwise not.}

\item{sorted}{If \code{TRUE}, elements are ordered by their names
before being compared, otherwise not.}

\item{...}{Not used.}
}
\value{
A list.
}
\description{
List representation of a list environment
}
\keyword{internal}
