% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listenv.R
\name{get_variable}
\alias{get_variable}
\alias{get_variable.listenv}
\title{Get name of variable for a specific element of list environment}
\usage{
get_variable(...)
}
\arguments{
\item{x}{A list environment.}

\item{name}{The name or index of element of interest.}
}
\value{
The name of the underlying variable
}
\description{
Get name of variable for a specific element of list environment
}
\keyword{internal}

