% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listenv.R
\name{listenv}
\alias{as.listenv}
\alias{listenv}
\title{Create a list environment}
\usage{
listenv(...)

as.listenv(...)
}
\arguments{
\item{\dots}{(optional) Named and/or unnamed objects to be
assigned to the list environment.}
}
\value{
An environment of class `listenv`.
}
\description{
Create a list environment
}
\examples{
x <- listenv(a=2, b=3, c="hello")
print(names(x))
names(x)[2] <- "B"
x$d <- 5:8

y <- as.list(x)
str(y)

z <- as.listenv(y)
}

