% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listenv.R
\name{listenv}
\alias{listenv}
\alias{as.listenv}
\alias{as.listenv}
\title{Create a list environment}
\usage{
listenv(...)

as.listenv(...)
}
\arguments{
\item{\dots}{(optional) Named and/or unnamed objects to be
assigned to the list environment.}
}
\value{
An environment of class `listenv`.
}
\description{
Create a list environment
}
\examples{
x <- listenv(c = 2, a = 3, d = "hello")
print(names(x))
names(x)[2] <- "A"
x$b <- 5:8

y <- as.list(x)
str(y)

z <- as.listenv(y)
}
