% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listenv.R
\name{listenv}
\alias{as.listenv}
\alias{listenv}
\title{Create a list environment}
\usage{
listenv(length = 0L)

as.listenv(...)
}
\arguments{
\item{length}{The number of NULL elements from start.}

\item{...}{The object to coerce and optional arguments.}
}
\value{
An environment of class `listenv`.
}
\description{
Create a list environment
}

