% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listenv.R
\name{listenv}
\alias{as.listenv}
\alias{listenv}
\title{Create a list environment}
\usage{
listenv(...)

as.listenv(...)
}
\arguments{
\item{...}{(optional) Named and/or unnamed objects to be
assigned to the list environment.}
}
\value{
An environment of class `listenv`.
}
\description{
Create a list environment
}

