% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listenv.R
\name{mapping}
\alias{mapping}
\alias{mapping.listenv}
\alias{map}
\title{Variable name map for elements of list environment}
\usage{
mapping(x, ...)

map(x, ...)
}
\arguments{
\item{x}{A list environment.}
}
\value{
A named character vector
}
\description{
Variable name map for elements of list environment
}
\details{
\emph{Functions \code{map()} and \code{map<-()} have been renamed to
\code{mapping()} and \code{mapping<-()}. The former will soon become
deprecated and eventually defunct. Please update accordingly.}
}
\keyword{internal}
