% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_env_subset.R
\name{parse_env_subset}
\alias{parse_env_subset}
\title{Helper function to infer target from expression and environment}
\usage{
parse_env_subset(expr, envir = parent.frame(), substitute = TRUE)
}
\arguments{
\item{expr}{An expression.}

\item{envir}{An environment.}

\item{substitute}{If TRUE, then the expression is
\code{substitute()}:ed, otherwise not.}
}
\value{
A named list.
}
\description{
Helper function to infer target from expression and environment
}
\keyword{internal}

